Make it work on Github pages
RooCode interaction:
--------------------

Prompt:
-------
when uploading this to github pages on
https://jheidbrink.github.io/browser-notifications-demo/, when I click
"send push notification" and click on the notification, a new browser
tab with url https://jheidbrink.github.io/ opens and returns a 404. Why
does a new browser tab open, and how is the URL for it calculated.
Please explain and fix that behaviour

Answer:
-------
I can explain the issue and fix it. Here's what's happening:

The Problem
Why a new tab opens: The service worker's notificationclick event handler is designed to either focus an existing tab or open a new one when a notification is clicked.

How the URL is calculated:

Line 79 and 181 in sw.js set url: '/' in the notification data
Line 117 uses const urlToOpen = notificationData.url || '/';
Line 133 calls clients.openWindow(urlToOpen) with '/'
The browser resolves '/' relative to the current origin, so on https://jheidbrink.github.io/browser-notifications-demo/, it becomes https://jheidbrink.github.io/ (the root domain)
Why it's 404: Because https://jheidbrink.github.io/ doesn't exist - your site is at https://jheidbrink.github.io/browser-notifications-demo/

The Fix
I need to change the URL to be relative to the current page location instead of the root domain.